+++ vizro-ai/examples/example.py
@@ -1,8 +1,9 @@
 """AI plot example."""
 
-import plotly.express as px
+import vizro.plotly.express as px
 from vizro_ai import VizroAI
 
 vizro_ai = VizroAI()
 df = px.data.gapminder()
-vizro_ai.plot(df, "describe the composition of gdp in continent,and horizontal line for avg gdp")
+fig = vizro_ai.plot(df, "describe the composition of gdp in continent,and horizontal line for avg gdp")
+fig.show()