+++ vizro-core/src/vizro/models/_components/_components_utils.py
@@ -1,4 +1,5 @@
 import logging
+import uuid
 from functools import partial
 
 from vizro.managers import data_manager
@@ -23,29 +24,30 @@ def _callable_mode_validator_factory(mode: str):
     return validator("figure", allow_reuse=True)(check_callable_mode)
 
 
-def _process_callable_data_frame(captured_callable, values):
+def _process_callable_data_frame(captured_callable):
+    # Possibly all this validator's functionality should move into CapturedCallable (or a subclass of it) in the
+    # future. This would mean that data is added to the data manager outside the context of a dashboard though,
+    # which might not be desirable.
     data_frame = captured_callable["data_frame"]
 
-    # Enable running e.g. px.scatter("iris") from the Python API and specification of "data_frame": "iris" through JSON.
-    # In these cases, data already exists in the data manager and just needs to be linked to the component.
     if isinstance(data_frame, str):
-        data_manager._add_component(values["id"], data_frame)
+        # Named data source, which could be dynamic or static. This means px.scatter("iris") from the Python API and
+        # specification of "data_frame": "iris" through JSON. In these cases, data already exists in the data manager.
         return captured_callable
 
-    # Standard case for px.scatter(df: pd.DataFrame).
-    # Extract dataframe from the captured function and put it into the data manager.
-    dataset_name = str(id(data_frame))
+    # Unnamed data source, which must be a pd.DataFrame and hence static data. This means px.scatter(pd.DataFrame())
+    # and is only possible from the Python API. Extract dataframe from the captured function and put it into the
+    # data manager.
+    # Unlike with model_manager, it doesn't matter if the random seed is different across workers here. So long as
+    # we always fetch static data from the data manager by going through the appropriate Figure component, the right
+    # data source name will be fetched. It also doesn't matter if multiple Figures with the same underlying data
+    # each have their own entry in the data manager, since the underlying pd.DataFrame will still be the same and
+    # not copied into each one, so no memory is wasted.
+    # Replace the "data_frame" argument in the captured callable with the data_source_name for consistency with
+    # dynamic data and to save memory. This way we always access data via the same interface regardless of whether it's
+    # static or dynamic.
+    data_source_name = str(uuid.uuid4())
+    data_manager[data_source_name] = data_frame
+    captured_callable["data_frame"] = data_source_name
 
-    logger.debug("Adding data to data manager for Figure with id %s", values["id"])
-    # If the dataset already exists in the data manager then it's not a problem, it just means that we don't need
-    # to duplicate it. Just log the exception for debugging purposes.
-    try:
-        data_manager[dataset_name] = data_frame
-    except ValueError as exc:
-        logger.debug(exc)
-
-    data_manager._add_component(values["id"], dataset_name)
-
-    # No need to keep the data in the captured function any more so remove it to save memory.
-    del captured_callable["data_frame"]
     return captured_callable