+++ vizro-core/tests/unit/vizro/models/_components/test_graph.py
@@ -78,9 +78,9 @@ class TestDunderMethodsGraph:
 
         assert graph.layout.margin.t == expected
         assert graph.layout.template.layout.margin.t == 64
-        assert graph.layout.template.layout.margin.l == 80
+        assert graph.layout.template.layout.margin.l == 24
         assert graph.layout.template.layout.margin.b == 64
-        assert graph.layout.template.layout.margin.r == 12
+        assert graph.layout.template.layout.margin.r == 24
 
     def test_update_theme_outside_callback(self, standard_px_chart):
         graph = vm.Graph(figure=standard_px_chart).__call__()
@@ -119,18 +119,15 @@ class TestAttributesGraph:
         assert "modelID" in graph._filter_interaction_input
 
 
-class TestProcessFigureDataFrame:
+class TestProcessGraphDataFrame:
     def test_process_figure_data_frame_str_df(self, standard_px_chart_with_str_dataframe, gapminder):
         data_manager["gapminder"] = gapminder
-        graph_with_str_df = vm.Graph(id="text_graph", figure=standard_px_chart_with_str_dataframe)
-        assert data_manager._get_component_data("text_graph").equals(gapminder)
-        assert graph_with_str_df["data_frame"] == "gapminder"
+        graph = vm.Graph(id="graph", figure=standard_px_chart_with_str_dataframe)
+        assert data_manager[graph["data_frame"]].load().equals(gapminder)
 
     def test_process_figure_data_frame_df(self, standard_px_chart, gapminder):
-        graph_with_df = vm.Graph(id="text_graph", figure=standard_px_chart)
-        assert data_manager._get_component_data("text_graph").equals(gapminder)
-        with pytest.raises(KeyError, match="'data_frame'"):
-            graph_with_df.figure["data_frame"]
+        graph = vm.Graph(id="graph", figure=standard_px_chart)
+        assert data_manager[graph["data_frame"]].load().equals(gapminder)
 
 
 class TestBuild: