+++ vizro-core/tests/unit/vizro/tables/test_dash_ag_grid.py
@@ -3,6 +3,7 @@ import pandas as pd
 import vizro.models as vm
 from asserts import assert_component_equal
 from dash import dcc, html
+from pandas import Timestamp
 from vizro.models.types import capture
 from vizro.tables import dash_ag_grid
 
@@ -14,6 +15,24 @@ data = pd.DataFrame(
         "date": pd.to_datetime(["2021/01/01", "2021/01/02", "2021/01/03"]),
     }
 )
+column_defs = [{"field": "cat"}, {"field": "int"}, {"field": "float"}, {"field": "date"}]
+row_data_date_converted = [
+    {"cat": "a", "int": 4, "float": 7.3, "date": "2021-01-01"},
+    {"cat": "b", "int": 5, "float": 8.2, "date": "2021-01-02"},
+    {"cat": "c", "int": 6, "float": 9.1, "date": "2021-01-03"},
+]
+row_data_date_raw = [
+    {"cat": "a", "date": Timestamp("2021-01-01 00:00:00"), "float": 7.3, "int": 4},
+    {"cat": "b", "date": Timestamp("2021-01-02 00:00:00"), "float": 8.2, "int": 5},
+    {"cat": "c", "date": Timestamp("2021-01-03 00:00:00"), "float": 9.1, "int": 6},
+]
+default_col_defs = {
+    "filter": True,
+    "filterParams": {"buttons": ["apply", "reset"], "closeOnApply": True},
+    "resizable": True,
+    "sortable": True,
+}
+style = {"height": "100%"}
 
 
 class TestDashAgGrid:
@@ -22,17 +41,14 @@ class TestDashAgGrid:
         assert_component_equal(
             grid,
             dag.AgGrid(
-                columnDefs=[{"field": "cat"}, {"field": "int"}, {"field": "float"}, {"field": "date"}],
-                rowData=[
-                    {"cat": "a", "int": 4, "float": 7.3, "date": "2021-01-01"},
-                    {"cat": "b", "int": 5, "float": 8.2, "date": "2021-01-02"},
-                    {"cat": "c", "int": 6, "float": 9.1, "date": "2021-01-03"},
-                ],
+                columnDefs=column_defs,
+                rowData=row_data_date_converted,
+                defaultColDef=default_col_defs,
+                style=style,
             ),
-            keys_to_strip={"className", "defaultColDef", "dashGridOptions", "style"},
+            keys_to_strip={"className", "dashGridOptions"},
         )
-        # we could test other properties such as defaultColDef,
-        # but this would just test our chosen defaults, and no functionality really
+        # skipping only dashGridOptions as this is mostly our defaults for data formats, and would crowd the tests
 
 
 class TestCustomDashAgGrid:
@@ -58,12 +74,48 @@ class TestCustomDashAgGrid:
             [
                 None,
                 html.Div(
-                    dag.AgGrid(id="__input_custom_ag_grid", columnDefs=[], rowData=[]),
+                    dag.AgGrid(id="__input_custom_ag_grid", columnDefs=column_defs, rowData=row_data_date_raw),
                     id=id,
                     className="table-container",
                 ),
             ],
-            id=f"{id}_outer",
+            color="grey",
+            parent_className="loading-container",
+        )
+
+        assert_component_equal(custom_grid, expected_grid)
+
+    def test_custom_dash_ag_grid_column_referral(self):
+        """Tests whether a custom created grid can be correctly built in vm.AgGrid.
+
+        This test focuses on the case that the custom grid includes column referrals on presumed data knowledge.
+        """
+        id = "custom_ag_grid"
+
+        @capture("ag_grid")
+        def custom_ag_grid(data_frame):
+            data_frame["cat"]  # access "existing" column
+            return dag.AgGrid(
+                columnDefs=[{"field": col} for col in data_frame.columns],
+                rowData=data_frame.to_dict("records"),
+            )
+
+        grid_model = vm.AgGrid(
+            id=id,
+            figure=custom_ag_grid(data_frame=data),
+        )
+        grid_model.pre_build()
+        custom_grid = grid_model.build()
+
+        expected_grid = dcc.Loading(
+            [
+                None,
+                html.Div(
+                    dag.AgGrid(id="__input_custom_ag_grid", columnDefs=column_defs, rowData=row_data_date_raw),
+                    id=id,
+                    className="table-container",
+                ),
+            ],
             color="grey",
             parent_className="loading-container",
         )