+++ vizro-core/CHANGELOG.md
@@ -11,6 +11,38 @@ See the fragment files in the [changelog.d directory](https://github.com/mckinse
 
 <!-- scriv-insert-here -->
 
+<a id='changelog-0.1.16'></a>
+
+# 0.1.16 — 2024-04-30
+
+## Removed
+
+- Remove class names `checkboxes-list`, `radio-items-list` and `input-container` from Vizro stylesheets. ([#414](https://github.com/mckinsey/vizro/pull/414))
+
+## Changed
+
+- Update design of dashboard layout. ([#433](https://github.com/mckinsey/vizro/pull/433))
+
+<a id='changelog-0.1.15'></a>
+
+# 0.1.15 — 2024-04-15
+
+## Highlights ✨
+
+- Add dynamic data, which can be reloaded while the dashboard is running. An optional caching layer enables efficient refreshes with per-data source timeouts. Visit the [user guide on data](https://vizro.readthedocs.io/en/stable/pages/user-guides/data/) for more details. ([#398](https://github.com/mckinsey/vizro/pull/398))
+
+## Changed
+
+- Replace default bootstrap stylesheet with `vizro-bootstrap` stylesheet. ([#384](https://github.com/mckinsey/vizro/pull/384))
+
+- Refactor code and remove custom classNames from `Button`, `Card`, `NavBar` and `NavLink`. ([#384](https://github.com/mckinsey/vizro/pull/384))
+
+- Change default continuous color scale to `SEQUENTIAL_CYAN`. ([#407](https://github.com/mckinsey/vizro/pull/407))
+
+## Fixed
+
+- Fix CSS for `floatingFilter` in `AgGrid`. ([#388](https://github.com/mckinsey/vizro/pull/388))
+
 <a id='changelog-0.1.14'></a>
 
 # 0.1.14 — 2024-03-26